# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/login": {
            "post": {
                "description": "Login",
                "summary": "Login",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{\"username\":\"ckman\",\"token\":\"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9\"}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Logout",
                "summary": "Logout",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"success\",\"data\":nil}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/package": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get package list",
                "summary": "Get package list",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":[\"20.8.5.45\"]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload package",
                "consumes": [
                    "multipart/form-data"
                ],
                "summary": "Upload package",
                "parameters": [
                    {
                        "type": "file",
                        "description": "package",
                        "name": "package",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"success\",\"data\":null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete package",
                "summary": "Delete package",
                "parameters": [
                    {
                        "type": "string",
                        "default": "20.8.5.45",
                        "description": "package version",
                        "name": "packageVersion",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"success\",\"data\":null}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/zk/replicated_table/{clusterName}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get replicated table in Zookeeper status",
                "summary": "Get replicated table in  Zookeeper status",
                "parameters": [
                    {
                        "type": "string",
                        "default": "test",
                        "description": "cluster name",
                        "name": "clusterName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{\"header\":[[\"vm101106\",\"vm101108\"],[\"vm102114\",\"vm101110\"],[\"vm102116\",\"vm102115\"]],\"tables\":[{\"name\":\"sensor_dt_result_online\",\"values\":[[\"l1846\",\"f1846\"],[\"l1845\",\"f1845\"],[\"l1846\",\"f1846\"]]}]}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/zk/status/{clusterName}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Zookeeper cluster status",
                "summary": "Get Zookeeper cluster status",
                "parameters": [
                    {
                        "type": "string",
                        "default": "test",
                        "description": "cluster name",
                        "name": "clusterName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":[{\"host\":\"192.168.102.116\",\"version\":\"3.6.2\",\"server_state\":\"follower\",\"peer_state\":\"following - broadcast\",\"avg_latency\":0.4929,\"approximate_data_size\":141979,\"znode_count\":926}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.LoginReq": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "63cb91a2ceb9d4f7c8b1ba5e50046f52"
                },
                "username": {
                    "type": "string",
                    "example": "ckman"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}